Added User Service
The service was implemented using MongoTemplate, which allows for
advanced database operations. The class includes functionality to pull a
single instance of user, as well as list, along with save, update &
delete. In addition, an extra function was added to flexibily toggle the
status of the user - inactive or not.